requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
